Added Exceptions and class Status changed

+++ src/main/java/eu/execom/pomodoro/service/PomodoroService.java
@@ -5,6 +5,7 @@ import eu.execom.pomodoro.repository.PomodoroRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.EntityExistsException;
 import java.util.List;
 
 @Service
@@ -13,19 +14,26 @@ public class PomodoroService {
     @Autowired
     private PomodoroRepository pomodoroRepository;
 
+    @Autowired
+    private Pomodoro savedPomodoro;
+
     public List<Pomodoro> getAll() { return pomodoroRepository.findAll(); }
 
     public Pomodoro getById(Long id) { return pomodoroRepository.getOne(id); }
 
     public Pomodoro save(Pomodoro pomodoro) {
-        Pomodoro savedPomodoro = pomodoroRepository.save(pomodoro);
+        try {
+            savedPomodoro = pomodoroRepository.save(pomodoro);
+        } catch (EntityExistsException e){
+            System.out.println("This entity already exists in database.");
+        }
         return savedPomodoro;
     }
 
     public void delete(Long id) {
-//        if (!pomodoroRepository.existsById(id)) {
-//
-//        }
+        if (!pomodoroRepository.existsById(id)) {
+            throw new EntityExistsException("Pomodoro with this id doesn't exist in database.");
+        }
         pomodoroRepository.deleteById(id);
     }
 }
