+++ src/main/java/eu/execom/pomodoro/service/TeamService.java
@@ -5,6 +5,7 @@ import eu.execom.pomodoro.repository.TeamRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.EntityExistsException;
 import java.util.List;
 
 @Service
@@ -12,6 +13,9 @@ public class TeamService {
 
     @Autowired
     private TeamRepository teamRepository;
+
+    @Autowired
+    private Team savedTeam;
 
     public List<Team> getAll() {
         return teamRepository.findAll();
@@ -22,11 +26,18 @@ public class TeamService {
     }
 
     public Team save(Team team) {
-        Team savedTeam = teamRepository.save(team);
+        try {
+            savedTeam = teamRepository.save(team);
+        } catch (EntityExistsException e){
+            System.out.println("This entity already exists in database.");
+        }
         return savedTeam;
     }
 
     public void delete(Long id) {
+        if (!teamRepository.existsById(id)) {
+            throw new EntityExistsException("Team with this id doesn't exist in database.");
+        }
         teamRepository.deleteById(id);
     }
